moved custom commands hanlder to controls.ts
