+++ src/react/KeybindingsScreenProvider.tsx
@@ -1,24 +1,8 @@
 import { createContext, useState, useEffect } from 'react'
-import { useSnapshot } from 'valtio'
 import { contro } from '../controls'
-import { customCommandsConfig } from '../customCommands'
-import { miscUiState } from '../globalState'
 import KeybindingsScreen from './KeybindingsScreen'
 import { useIsModalActive } from './utilsApp'
-import { CustomCommand } from './KeybindingsCustom'
 
-
-const customCommandsHandler = (buttonData: { code?: string, button?: string, state: boolean }) => {
-  if (!buttonData.state) return
-
-  const codeOrButton = buttonData.code ?? buttonData.button
-  const inputType = buttonData.code ? 'keys' : 'gamepad'
-  for (const value of Object.values(contro.userConfig!.custom)) {
-    if (value[inputType]?.includes(codeOrButton!)) {
-      customCommandsConfig[(value as CustomCommand).type].handler((value as CustomCommand).inputs)
-    }
-  }
-}
 
 export const updateBinds = (commands: any) => {
   contro.inputSchema.commands.custom = Object.fromEntries(Object.entries(commands?.custom ?? {}).map(([key, value]) => {
@@ -45,14 +29,6 @@ export const updateBinds = (commands: any) => {
       return [key, newValue]
     }))
   }
-
-  if (!commands['custom']) return
-
-  //todo: don't trigger handler when setting new binding
-  contro.enabled = false
-  contro.off('pressedKeyOrButtonChanged', customCommandsHandler)
-  contro.on('pressedKeyOrButtonChanged', customCommandsHandler)
-  contro.enabled = true
 }
 
 const bindingActions = {
@@ -64,15 +40,6 @@ export const BindingActionsContext = createContext(bindingActions)
 export default () => {
   const [bindActions, setBindActions] = useState(bindingActions)
   const isModalActive = useIsModalActive('keybindings')
-  const { gameLoaded } = useSnapshot(miscUiState)
-
-  useEffect(() => {
-    if (gameLoaded) {
-      contro.on('pressedKeyOrButtonChanged', customCommandsHandler)
-    } else {
-      contro.off('pressedKeyOrButtonChanged', customCommandsHandler)
-    }
-  }, [gameLoaded])
 
   if (!isModalActive) return null
